feat(search): implement CollectionsFilter component with optimistic UI update and concurrent navigation using startTransition to avoid UI blocking during router push
