+++ src/app/shop/layout.tsx
@@ -1,7 +1,7 @@
 import { getWixServerClient } from "@/lib/wix-server.base";
 import { getCollections } from "@/wix-api/collections";
 import React from "react";
-import SearchFilterLayout from "./components/SearchFilterLayout";
+import SearchFilterLayout from "./components/searchFilterLayout/SearchFilterLayout";
 
 interface LayoutProps {
   children: React.ReactNode;
